fix(parquet): isolate wasm worker dependencies
